% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.R
\name{calculate_total_acres_burned}
\alias{calculate_total_acres_burned}
\title{Calculate Total Acres Burned}
\usage{
calculate_total_acres_burned()
}
\value{
An integer representing the total acres burned.
}
\description{
Calculates the total number of acres burned in all recorded incidents.
}
\examples{
calculate_total_acres_burned()
}
