% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\name{fire}
\alias{fire}
\title{Forest Fire Incidents}
\format{
A data frame with 4 columns:
\describe{
  \item{Date}{Date of the incident.}
  \item{Location}{Location of the forest fire.}
  \item{Severity}{Severity of the fire, categorized as 'High', 'Medium', or 'Low'.}
  \item{AcresBurned}{The number of acres burned in the fire.}
}
}
\usage{
fire()
}
\description{
A dataset containing information about forest fire incidents.
}
