% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\docType{data}
\name{fire_data}
\alias{fire_data}
\title{Fire Incident Data}
\format{
A data frame with 4 rows and 5 columns:
\describe{
  \item{\code{IncidentID}}{An integer vector of incident IDs.}
  \item{\code{Date}}{The dates of the incidents as \code{Date} objects.}
  \item{\code{Location}}{Character vector of the locations of the incidents.}
  \item{\code{Severity}}{Character vector indicating the severity of each incident (High, Medium, Low).}
  \item{\code{AcresBurned}}{A numeric vector of the number of acres burned in each incident.}
}
}
\usage{
fire_data
}
\description{
A dataset containing information on various fire incidents. Each row represents a unique fire incident with details about the date, location, severity, and acres burned.
}
\keyword{datasets}
