% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{get_fire_details_by_id}
\alias{get_fire_details_by_id}
\title{Get Fire Details by Incident ID}
\usage{
get_fire_details_by_id(incident_id)
}
\arguments{
\item{incident_id}{An integer representing the Incident ID.}
}
\value{
A data frame row with the details of the specified fire incident.
}
\description{
Retrieves details of a forest fire incident given its ID.
}
\examples{
get_fire_details_by_id(2)
}
