% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f4.R
\name{list_incidents_by_location}
\alias{list_incidents_by_location}
\title{List Incidents by Location}
\usage{
list_incidents_by_location(location)
}
\arguments{
\item{location}{A string representing the location.}
}
\value{
A subset of the fire_data data frame.
}
\description{
Lists all fire incidents recorded at a given location.
}
\examples{
list_incidents_by_location("ForestA")
}
