% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_two.R
\name{mean_of_two}
\alias{mean_of_two}
\title{Calculate the mean of two numbers}
\usage{
mean_of_two(a, b)
}
\arguments{
\item{a}{First number}

\item{b}{Second number}
}
\value{
Mean of a and b
}
\description{
Calculate the mean of two numbers
}
