% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab8.R, R/myboot2.R, R/mychisim.R, R/myci.R,
%   R/myf.R, R/myfun.R, R/mymlbeta.R
\name{myf}
\alias{myf}
\title{Title of your function}
\usage{
myf(iter = 100, n = 10, p = 0.5)

myf(iter = 100, n = 10, p = 0.5)

myf(iter = 100, n = 10, p = 0.5)

myf(iter = 100, n = 10, p = 0.5)

myf(iter = 100, n = 10, p = 0.5)

myf(iter = 100, n = 10, p = 0.5)

myf(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{A brief description of this parameter.}

\item{n}{A brief description of this parameter.}

\item{p}{A brief description of this parameter.}
}
\value{
Describe what your function returns.

Describe what your function returns.

Describe what your function returns.

Describe what your function returns.

Describe what your function returns.

A vector of squared components

Describe what your function returns.
}
\description{
A detailed description of what your function does.

A detailed description of what your function does.

A detailed description of what your function does.

A detailed description of what your function does.

A detailed description of what your function does.

A detailed description of what your function does.
}
\examples{
myf(iter=100, n=10, p=0.5)
myf(iter=100, n=10, p=0.5)
myf(iter=100, n=10, p=0.5)
myf(iter=100, n=10, p=0.5)
myf(iter=100, n=10, p=0.5)
myf(iter=100, n=10, p=0.5)
}
