% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot and Calculate Area Under Normal Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sigma}{Standard deviation of the normal distribution.}

\item{a}{The x-value up to which the area under the curve is shaded and probability calculated.}
}
\value{
A list containing the calculated probability (area under the curve) up to the value `a`.
}
\description{
This function plots a normal distribution curve with given mean (`mu`) and standard deviation (`sigma`),
shades the area under the curve up to a specified value (`a`), and calculates the cumulative probability up to that point.
}
\examples{
myncurve(mu = 0, sigma = 1, a = 1.96)
}
